+++ src/components/Message/Message.module.css
@@ -12,26 +12,3 @@
   color: white;
   font-weight: normal;
 }
-.form {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 15px;
-}
-.form #message_input {
-  width: 100%;
-  border-radius: 5px 0 0 5px;
-  padding: 10px;
-  border: 2px solid #d3d3d3;
-  font-size: 18px;
-}
-.form #message_input:focus,
-.form #message_input:active {
-  outline: none;
-}
-.form #message_send {
-  border-radius: 0 5px 5px 0;
-  padding: 10px 50px;
-  border: 2px solid #d3d3d3;
-  background-color: #d3d3d3;
-  font-size: 18px;
-}
